refactor(internal-account): split server.go into domain endpoint files
Extract RPC handlers from the monolithic server.go (964 lines) into
domain-grouped files following the service-file-conventions guide:

- grpc_account_endpoints.go: Initiate, Update, Control + saveAccountWithEvent
- grpc_balance_endpoints.go: GetBalance + mapPositionKeepingErrorToGRPC
- grpc_query_endpoints.go: Retrieve, List

server.go retains struct, interfaces, constructors, options, constants,
maps, and shared helpers (findAccountByID). Now 266 lines.

Remove from knownOversizedFiles in architecture test.
