+++ tests/architecture/size_test.go
@@ -29,7 +29,6 @@ var knownOversizedFiles = map[string]bool{
 	"services/control-plane/internal/generator/validate_fix.go":             true,
 	"services/current-account/service/saga_handlers.go":                     true,
 	"services/identity/service/server.go":                                   true,
-	"services/internal-account/service/server.go":                           true,
 	"services/market-information/service/observation_service.go":            true,
 	"services/mcp-server/internal/tools/economy.go":                         true,
 	"services/mcp-server/internal/tools/refdata.go":                         true,
